requirements.txt
streamlit
pdfplumber
pdf2image
pytesseract
Pillow